requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Bcrypt==1.0.1
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
propcache==0.3.1
PyJWT==2.10.1
pytesseract==0.3.13
requests==2.32.3
twilio==9.5.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
